server/requirements.txt
asgiref==3.8.1
attrs==24.2.0
backports.zoneinfo==0.2.1
certifi==2024.7.4
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
Django==4.2.15
django-cors-headers==4.4.0
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
idna==3.8
importlib-resources==6.4.4
inflection==0.5.1
itypes==1.2.0
jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.5
openapi-codec==1.3.2
pillow==10.4.0
pkgutil-resolve-name==1.3.10
PyJWT==2.9.0
python-dateutil==2.9.0.post0
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rpds-py==0.20.0
simplejson==3.19.3
six==1.16.0
sqlparse==0.5.1
typing-extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
whitenoise==6.7.0
zipp==3.20.0
asgiref==3.8.1
attrs==24.2.0
backports.zoneinfo==0.2.1
certifi==2024.7.4
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
Django==4.2.15
django-cors-headers==4.4.0
django-environ==0.11.2
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
idna==3.8
importlib-resources==6.4.4
inflection==0.5.1
itypes==1.2.0
jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.5
openapi-codec==1.3.2
pillow==10.4.0
pkgutil-resolve-name==1.3.10
psycopg==3.2.2
psycopg2-binary==2.9.9
PyJWT==2.9.0
python-dateutil==2.9.0.post0
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rpds-py==0.20.0
simplejson==3.19.3
six==1.16.0
sqlparse==0.5.1
typing-extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
whitenoise==6.7.0
zipp==3.20.0
